# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/v1/customers/{id}/credit-adjustments": {
      "post": {
        "operationId": "CreditsController_request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditAdjustmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Credits"
        ]
      }
    },
    "/admin/v1/credit-adjustments": {
      "get": {
        "operationId": "CreditsController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING_APPROVAL",
                "APPROVED",
                "APPLIED",
                "REJECTED",
                "FAILED"
              ]
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "data": [],
                  "hasMore": false,
                  "nextCursor": null
                }
              }
            }
          }
        },
        "tags": [
          "Credits"
        ]
      }
    },
    "/admin/v1/credit-adjustments/{id}/approve": {
      "post": {
        "operationId": "CreditsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Credits"
        ]
      }
    },
    "/admin/v1/credit-adjustments/{id}/reject": {
      "post": {
        "operationId": "CreditsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Credits"
        ]
      }
    },
    "/admin/v1/credit-adjustments/{id}/retry": {
      "post": {
        "operationId": "CreditsController_retry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Credits"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zapiack Admin Console API",
    "description": "Backend for the Zapiack Admin Console.\n\n**Authentication is two steps.** `POST /auth/login` returns `totp_required` and\nsets a session cookie that every guarded route rejects until\n`POST /auth/2fa/verify` succeeds. Send `credentials: \"include\"` on every request.\n\n**Writes need `x-csrf-token`**, echoing the value from `/auth/2fa/verify` or\n`/auth/me`. Some also need an `Idempotency-Key` header and a written `reason`;\neach operation says which.\n\n**Every list returns `{ data, hasMore, nextCursor }`.** Extra context sits beside\n`data`, never instead of it.\n\n**Money is a decimal string** (`\"125000.0000\"`), never a number. Dates are\nAfrica/Lagos calendar days. Unknown query or body fields are rejected with 400.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zpk_admin_at",
        "description": "Session cookie, set by the login flow. Not readable from JavaScript."
      }
    },
    "schemas": {
...
      "ReasonDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 8,
            "maxLength": 1000
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "CreditAdjustmentDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 8,
            "maxLength": 1000
          },
          "direction": {
            "type": "string",
            "enum": [
              "CREDIT",
              "DEBIT"
            ]
          },
          "amountNgn": {
            "type": "string",
            "description": "Naira as a decimal string; money never travels as a float."
          }
        },
        "required": [
          "reason",
          "direction",
          "amountNgn"
        ]
      },
...
    }
  }
}